% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{array_gpcc}
\alias{array_gpcc}
\title{How to open all files together from a startpoint to an endpoint}
\usage{
array_gpcc(dir_GPCC, varid, start_date = NULL, end_date = NULL)
}
\arguments{
\item{dir_GPCC}{directory of the Global Precipitation Climatology Centre that contain the files of the variables to be downloaded}

\item{varid}{variable to be downloaded. This can take the values within c("", "", "")}

\item{start_date}{initial year of the subset. If NULL, the subset starts in 2013}

\item{end_date}{final year of the subset. If NULL, the subset ends in the current year}
}
\value{
this function returns a 3D array for the selected variable of the GPCC
}
\description{
This function reads and merge all files of the selected dir_GPCC
}
