% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{func_files_GPCC}
\alias{func_files_GPCC}
\title{How to get the files of a folder of the GPCC (Global Precipitation Climatology Centre)}
\usage{
func_files_GPCC(dir_GPCC)
}
\arguments{
\item{dir_GPCC}{directory of the Global Precipitation Climatology Centre that contain the files of the variables to be downloaded}
}
\value{
this function returns a list of paths of the selected variable of the GPCC
}
\description{
Function to get the list of files of GPCC (Global Precipitation Climatology Centre)
}
