% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{func_GPCC}
\alias{func_GPCC}
\title{Function to read a file of GPCC (Global Precipitation Climatology Centre)}
\usage{
func_GPCC(url, varid)
}
\arguments{
\item{url}{path of the file to be loaded}

\item{varid}{variable to be downloaded}
}
\value{
3D matrix
}
\description{
Function to read a file of GPCC (Global Precipitation Climatology Centre)
}
