% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broker.R
\name{getBroker}
\alias{getBroker}
\title{Factory method to get a stateful controller (broker of data).}
\usage{
getBroker()
}
\value{
the controller object
}
\description{
The factory creates a
controller object. It has state (selection of LTER site and Essential Variable). Broker of data.

The object exposes the methods:

setSite: set selected site (via its deimsUUID. It must be present in the sites table)
getSite: return the currently selected site record as a \code{tibble} with one row
siteList: return th named \code{list} of site deimsUUID, useful for selectizeInput in shiny UI
EVsList: return the named \code{list} of (current site's) EV ids, useful for selectizeInput in shiny UI
siteNames: return a \code{character} array with the available site names
evNames

getCurrentEVs: get the available EVs for the currently selected site (they depend on the domain of the site)
setEv: Set the selected EV. Warn if the \code{id} is not present in the object returned by \code{getCurrentEVs}
getEv: return the currently selected EV record as a tibble with one row

getInfo_Site:

getOtherResData:
getOtherRepoData:
getEVsData:
}
\author{
Paolo Tagliolato (ptagliolato)

Alessandro Oggioni (oggioniale)
}
