% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{runShinyApp}
\alias{runShinyApp}
\title{Run the main app}
\usage{
runShinyApp(...)
}
\arguments{
\item{...}{other params passed to \code{shiny::runApp}}
}
\description{
Run the main app
}
\examples{
ITINERIS.EVsVRE::runShinyApp(launch.browser=rstudioapi::viewer)
}
